change gyro & accel dt from float to uint64. This has the benefit of calculating the estimator timeslip correctly.

+++ src/modules/sdlog2/sdlog2_messages.h
@@ -518,8 +518,8 @@ struct log_EST5_s {
 #define LOG_RPL1_MSG 51
 struct log_RPL1_s {
 	uint64_t time_ref;
-	float gyro_integral_dt;
-	float accelerometer_integral_dt;
+	uint64_t gyro_integral_dt;
+	uint64_t accelerometer_integral_dt;
 	uint64_t magnetometer_timestamp;
 	uint64_t baro_timestamp;
 	float gyro_x_rad;
